+++ src/config.ts
@@ -1,3 +1,4 @@
 export const config = {
   debug: false,
+  logger: console.log as (o: object) => string,
 };