+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.4.0
 
 - **_Breaking Change_**: set stdout to empty string when ripgrep no match, instead of throwing errors.
+- Feat: customizable logger.
 
 ## v0.3.1